Changin animations
Changin animations for a better functioning in intersection observer api
