+++ index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge" >
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>My portfolio</title>
   <!--Google fonts-->
@@ -50,11 +50,11 @@
       <div class="contenedor">
         <div class="row">
           <!--Seccion izquierda-->
-          <div class="col-md-6 jumbotron_section-left element-from-left">
-            <h1 class="subtitle">I'm
+          <div class="col-md-6 jumbotron_section-left">
+            <h1 class="subtitle element-from-left">I'm
               <hr /> <span class="title color-orange">Luis</span>
             </h1>
-            <article class="social-media">
+            <article class="social-media element-from-left">
               <a href="https://www.linkedin.com/in/luis-angel-zepeda-gonzalez" target="_blank"
                 class="social-media_item">
                 <i class="fa-brands fa-linkedin"></i>
@@ -66,29 +66,24 @@
             </article>
           </div>
           <!--Seccion derecha-->
-          <div class="col-md-6 jumbotron_section-right element-from-right">
-            <p class="subsection_title color-orange fw-bolder">Full Stack
+          <div class="col-md-6 jumbotron_section-right">
+            <p class="subsection_title color-orange fw-bolder element-from-right">Full Stack
               <br />
               <span class="color-white">Web Developer <br />Java Script</span>
             </p>
-            <a href="./assets/archivos/cv_Luis-Angel-Zepeda-Gonzalez.pdf" class="btn-one" target="_blank">DownLoad CV</a>
+            <a href="./assets/archivos/cv_Luis-Angel-Zepeda-Gonzalez.pdf" class="btn-one element-from-right" target="_blank">DownLoad CV</a>
           </div>
-          <div class="jumbotron-circle element-from-grow">
+          <div class="jumbotron-circle">
           </div>
         </div>
       </div>
-      <!-- <section class="burbujas">
-        <div class="burbuja"></div>
-        <div class="burbuja"></div>
-        <div class="burbuja"></div>
-        <div class="burbuja"></div>
-      </section> -->
+      
     </section>
     <!--About me-->
-    <section class="about-me element-from-up bg-black" id="about-me">
-      <div class="contenedor">
+    <section class="about-me bg-black" id="about-me">
+      <div class="contenedor element-from-left">
         <h2 class="subsection_title color-black">About me</h2>
-        <p class="about-me_text color-white">
+        <p class="about-me_text color-white ">
           I'm a full stack developer in progress at Academlo bootcamp.
           I currently have experience on the front end side, I'm passionate about everything related to technology.
           My favorite hobbies are going to the beach, playing video games and listening to music.
@@ -99,7 +94,7 @@
       </div>
     </section>
     <!--Proyects-->
-    <section class="proyects element-from-up bg-green" id="proyects">
+    <section class="proyects bg-green" id="proyects">
 
       <div class="contenedor">
         <h2 class="subsection_title color-white text-center">
@@ -167,7 +162,7 @@
       </div>
     </section>
     <!--Tecnologies-->
-    <section class="tecnologies element-from-up bg-black" id="tecnologies">
+    <section class="tecnologies bg-black" id="tecnologies">
       <div class="contenedor">
         <h2 class="subsection_title color-black text-center">
           Tecnologies
@@ -229,15 +224,15 @@
             <i class="fw-bolder bg-green color-sequelize">
               <img src="./assets/images/express-logo.svg" alt="express-logo">
               <p>Sequelize</p>
-          </i>
-            
+            </i>
+
           </li>
         </ul>
 
       </div>
     </section>
     <!--Contact-->
-    <section class="contact element-from-up bg-green" id="contact">
+    <section class="contact bg-green" id="contact">
       <div class="contenedor ">
         <h2 class="subsection_title color-white text-center">Contact me</h2>
         <div class="row">
@@ -261,23 +256,23 @@
           <div class="contact-item zoom right-item col-md-6">
             <div class="card text-center primary-shadow bg-white">
               <a href="mailto:luisangeluis@gmail.com?">
-              <svg width="40" height="40" viewBox="0 0 40 40" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <path
-                  d="M6.66683 6.66666H33.3335C35.1668 6.66666 36.6668 8.16666 36.6668 9.99999V30C36.6668 31.8333 35.1668 33.3333 33.3335 33.3333H6.66683C4.8335 33.3333 3.3335 31.8333 3.3335 30V9.99999C3.3335 8.16666 4.8335 6.66666 6.66683 6.66666Z"
-                  fill="#F26440" stroke="white" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" />
-                <path d="M36.6668 9.99999L20.0002 21.6667L3.3335 9.99999" fill="#F26440" />
-                <path d="M36.6668 9.99999L20.0002 21.6667L3.3335 9.99999" stroke="white" stroke-width="2"
-                  stroke-linecap="round" stroke-linejoin="round" />
-              </svg>
-              <div class="card-body">
-                <card class="card-title">
-                  Email
-                </card>
-                <div class="card-text">
-                  luisangeluis@gmail.com
+                <svg width="40" height="40" viewBox="0 0 40 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+                  <path
+                    d="M6.66683 6.66666H33.3335C35.1668 6.66666 36.6668 8.16666 36.6668 9.99999V30C36.6668 31.8333 35.1668 33.3333 33.3335 33.3333H6.66683C4.8335 33.3333 3.3335 31.8333 3.3335 30V9.99999C3.3335 8.16666 4.8335 6.66666 6.66683 6.66666Z"
+                    fill="#F26440" stroke="white" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" />
+                  <path d="M36.6668 9.99999L20.0002 21.6667L3.3335 9.99999" fill="#F26440" />
+                  <path d="M36.6668 9.99999L20.0002 21.6667L3.3335 9.99999" stroke="white" stroke-width="2"
+                    stroke-linecap="round" stroke-linejoin="round" />
+                </svg>
+                <div class="card-body">
+                  <card class="card-title">
+                    Email
+                  </card>
+                  <div class="card-text">
+                    luisangeluis@gmail.com
+                  </div>
                 </div>
-              </div>
-            </a>
+              </a>
             </div>
           </div>
         </div>
@@ -307,7 +302,7 @@
   </footer>
 
   <!-- Custom js -->
-  <script src="/assets/js/index.js"></script>
+  <script src="/assets/js/index.js" type="module"></script>
 </body>
 
 </html>
